updating tests and plotting function
